+++ CHANGELOG.md
@@ -10,6 +10,20 @@ You may also reference live-docs based on the openapi specification here:
 https://tapis-project.github.io/live-docs
 
 ---------------------------------------------------------------------------
+## 1.3.4 - 2023-06-14
+
+Incremental improvements and new features.
+
+### Breaking changes:
+- Environment variables beginning with *_tapis* may not be used in *jobEnvVariables*. This is to match Jobs service behavior.
+
+### New features:
+- Add notes and inputMode attributes to items in jobEnvVariables.
+
+### Bug fixes:
+- None
+
+---------------------------------------------------------------------------
 ## 1.3.3 - 2023-06-01
 
 Incremental improvements and bug fixes